+++ plugins/password/drivers/ldap_simple.php
@@ -119,11 +119,12 @@ class rcube_ldap_simple_password
         $this->debug = $rcmail->config->get('ldap_debug');
         $ldap_host   = $rcmail->config->get('password_ldap_host', 'localhost');
         $ldap_port   = $rcmail->config->get('password_ldap_port', '389');
+        $ldap_uri    = $this->_host2uri($ldap_host, $ldap_port);
 
-        $this->_debug("C: Connect to $ldap_host:$ldap_port");
+        $this->_debug("C: Connect [{$ldap_uri}]");
 
         // Connect
-        if (!$ds = ldap_connect($ldap_host, $ldap_port)) {
+        if (!($ds = ldap_connect($ldap_uri))) {
             $this->_debug("S: NOT OK");
 
             rcube::raise_error([
@@ -289,5 +290,25 @@ class rcube_ldap_simple_password
         if ($this->debug) {
             rcube::write_log('ldap', $str);
         }
+    }
+
+    /**
+     * Convert LDAP host/port into URI
+     */
+    private static function _host2uri($host, $port = null)
+    {
+        if (stripos($host, 'ldapi://') === 0) {
+            return $host;
+        }
+
+        if (strpos($host, '://') === false) {
+            $host = ($port == 636 ? 'ldaps' : 'ldap') . '://' . $host;
+        }
+
+        if ($port && !preg_match('/:[0-9]+$/', $host)) {
+            $host .= ':' . $port;
+        }
+
+        return $host;
     }
 }