+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix PHP8 deprecation warning in the reconnect plugin (#9083)
 - Fix "Show source" on mobile with x_frame_options = deny (#9084)
 - Fix various PHP warnings (#9098)
+- Fix deprecated use of ldap_connect() in password's ldap_simple driver (#9060)
 
 ## Release 1.6.2